+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+#1.0.2
+- Now link/image/text selection menus can be populated with regular menu actions
+- It is now possible to deactive link/image/text selection menu by removing all levels
+- Added tooltips in options
+
 #1.0.1
 - Added switch for debug mode in settings
 - Updated description